pyproject.toml
[project]
name = "gt-automacao-contratos"
version = "0.1.0"
description = ""
authors = [
    {name = "renansilva42",email = "renansilvageografia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.2,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "python-json-logger (>=3.2.1,<4.0.0)",
    "pydantic (>=2.10.6,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
